refactor: upcoming-performance 폴더명 변경
