+++ apps/client/src/pages/my/components/upcoming-performance/no-upcoming-performance-section.css.ts
