+++ apps/client/src/pages/my/components/upcoming-performance/upcoming-performance-section.tsx
